# specs/001-fullstack-todo-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Full-Stack Todo Application API
  version: 1.0.0
  description: API for managing user tasks with authentication.

servers:
  - url: http://localhost:8000/
    description: Local development server

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        message:
          type: string

paths:
  /auth/signup:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /auth/signin:
    post:
      summary: Authenticate user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Successful authentication, returns JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /api/{user_id}/tasks:
    get:
      summary: Retrieve all tasks for a specific user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the user to retrieve tasks for. Must match authenticated user's ID.
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - JWT token missing or invalid
        '403':
          description: Forbidden - user_id in path does not match authenticated user
    post:
      summary: Create a new task for a specific user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the user to create a task for. Must match authenticated user's ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input (e.g., empty title)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /api/{user_id}/tasks/{id}:
    get:
      summary: Retrieve a single task by ID for a specific user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the user who owns the task. Must match authenticated user's ID.
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the task to retrieve
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found for this user
    put:
      summary: Update an existing task by ID for a specific user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the user who owns the task. Must match authenticated user's ID.
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found for this user
    delete:
      summary: Delete a task by ID for a specific user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the user who owns the task. Must match authenticated user's ID.
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the task to delete
      responses:
        '204':
          description: Task successfully deleted (No Content)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found for this user

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle the completion status of a task by ID for a specific user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the user who owns the task. Must match authenticated user's ID.
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the task to mark as complete/incomplete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: Set to true to mark as complete, false to mark as incomplete.
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Task not found for this user